General cleanup of serializable support.
The following classes have been made serializable (either with
Serializable or with Externalizable, as appropriate):

 - Service
 - GroupConfiguraton
 - ProtocolFactory
 - Message
 - MembershipID
 - SocketAddress (implementations only)

This allows configurations to be stored persistently, for communication
state to be logged (including messages, view ids, and addresses), and
exchanged in messages in a protocol independent fashion.

In addition, all serialVersionUIDs of these classes have been set
to 2L (as some have been previously labeled as 1L).