Change pacakge structure.
First, move everything from the membership package to the main package. There is
no point in having a separate package for this when all protocols targetted by
jGCS will implement it.

Second, move abstract implementations from main package to a new spi package. This
mimics the structure of java.nio hierarchy, in which service provider infrastructure
is separated from the interfaces. This clearly identifies what interfaces end-users
should be concerned with.