Remove DataSession.send() and make use of Annotations.
Sending a message to a specific member of the group is offered by some
protocols, but not by all. Having this method encouraged non-portable
code.

The alternative is to consider message targetting as an optional
optimization. Applications should not depend on it, but can use
it while being portable.

The same thing for self-delivery (e.g. in Spread) that so far
was not supported.