General synchronization cleanup in spi.* classes.
A pair of related control and data sessions now uses a single shared
lock instead of synchronized blocks. This makes it easier to have shared
state without deadlocks. The new rules are:

 - Public methods implemented by Abstract*Sessions are thread-safe by
 using the lock.

 - Public methods implemented by derived concrete classes can be made
 thread safe also by using the lock, or by relying in the underlying
 protocol already being thread-safe itself.

 - Invocation of application callbacks, done through protected
 notify*Listener(...) methods, are thread-safe (will not NPE with
 concurrent set*Listener(null) or close()) but release the lock
 while invoking the application. They should be invoked only
 outside the lock.

 - Other protected utility methods are not thread-safe and should be
 used from concrete classes only when properly protected by the lock.

 - close() delegates to cleanup() in both sessions, invoked atomically
 in the contex of the lock. It should not block. Concrete classes
 should thus override cleanup(), call super.cleanup(), and do whatever
 cleanup is needed (i.e., disposing of the session).

This means that it is possible that some callbacks are invoked after
close() has ended. The application should deal with it. The alternative
would force close() to wait on the protocols and callback threads to
shutdown, potentially leading to the deadlock when close() is invoked
from a callback, which is likely to happen.