Removed ControlSession.isJoined, as it is inherently racy.
There is absolutely no guarantee that, whenever the application
reacts to the results, that it still is true, as it changes outside
the control of the application.

Applications should instead of using isJoined, try to do whathever
they want to do and deal with exceptions.

For protocols without membership, they should assume joined after
ControlSession.join() has completed. For protocols with membership,
it should be assumed only after the first membership notification.